Fix message_count type mismatch: PG INTEGER vs Rust i64
sightings.message_count is INTEGER (INT4) but HistoryRow expects i64
(INT8). Cast to BIGINT in the query to match.

+++ rust/adsb-server/src/db_pg.rs
@@ -634,7 +634,7 @@ impl AdsbDatabase for TimescaleDb {
         let rows = sqlx::query(
             "SELECT a.icao, s.callsign, a.country, a.is_military,
                     s.min_altitude_ft, s.max_altitude_ft,
-                    COALESCE(s.message_count, 0) as message_count,
+                    COALESCE(s.message_count, 0)::BIGINT as message_count,
                     EXTRACT(EPOCH FROM a.first_seen)::double precision as first_seen,
                     EXTRACT(EPOCH FROM a.last_seen)::double precision as last_seen
              FROM aircraft a
